[UI] Phase 5.a: Split dashboard resource config files by Kubernetes kind (#18660)
Splits three oversized dashboard resource column-definition files into one
file per Kubernetes kind so each falls below the 600-line warn threshold:

- configuration/config.tsx (1452 lines) -> 43 lines entry-point + 11 kind files
- workloads/config.tsx (1272 lines) -> 37 lines entry-point + 8 kind files
- network/config.tsx (916 lines) -> 33 lines entry-point + 6 kind files

The entry-point ConfigurationTableConfig / WorkloadTableConfig /
NetWorkTableConfig functions compose the same return object as before;
cell renderers stay inline inside each kind file (per plan §7.2 scope).

Removes the now-cleared entries from legacyMaxLineOffenders /
legacyMaxLineSoftOffenders in eslint.config.js, and adds service-columns.tsx
to legacyInlineStyleOffenders (the inline style={{ ... }} props moved out of
network/config.tsx into the new service kind file).

AUDIT size over_warn=13 over_hard=5 (down from over_warn=16 over_hard=7).

Skipping the pre-commit hook due to the documented Node 18 npx Node-binary
issue; lint-staged was run directly (node_modules/.bin/lint-staged) and
passed before this commit. npm run lint and npx tsc --noEmit both pass.

Refs #18660
Refs #18654
