+++ ui/components/dashboard/resources/configuration/config.tsx
@@ -1,14 +1,15 @@
-import React from 'react';
-import { timeAgo } from '../../../../utils/k8s-utils';
-import { SINGLE_VIEW } from '../config';
-import { Title } from '../../view';
-
-import { TooltipWrappedConnectionChip } from '../../../connections/ConnectionChip';
-import { DefaultTableCell, SortableTableCell } from '../sortable-table-cell';
-import { CONNECTION_KINDS } from '../../../../utils/Enum';
 import useKubernetesHook from '@/utils/hooks/useKubernetesHook';
-import { getK8sContextFromClusterId } from '../../../../utils/multi-ctx';
-import { FormatId } from '@/components/data-formatter';
+import { buildConfigMapColumns } from './configmap-columns';
+import { buildSecretColumns } from './secret-columns';
+import { buildResourceQuotaColumns } from './resource-quota-columns';
+import { buildLimitRangeColumns } from './limit-range-columns';
+import { buildHorizontalPodAutoscalerColumns } from './horizontal-pod-autoscaler-columns';
+import { buildVerticalPodAutoscalerColumns } from './vertical-pod-autoscaler-columns';
+import { buildPodDisruptionBudgetColumns } from './pod-disruption-budget-columns';
+import { buildPriorityClassColumns } from './priority-class-columns';
+import { buildRuntimeClassColumns } from './runtime-class-columns';
+import { buildLeasesColumns } from './leases-columns';
+import { buildMutatingWebhookConfigurationColumns } from './mutating-webhook-configuration-columns';
 
 export const ConfigurationTableConfig = (
   switchView,
@@ -18,1435 +19,25 @@ export const ConfigurationTableConfig = (
   workloadType,
 ) => {
   const ping = useKubernetesHook();
+  const args = {
+    switchView,
+    meshSyncResources,
+    k8sConfig,
+    connectionMetadataState,
+    workloadType,
+    ping,
+  };
   return {
-    ConfigMap: {
-      name: 'ConfigMap',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['data', 's'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'data',
-          label: 'Keys',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              if (!val) return <>-</>;
-              const parseVal = JSON.parse(val);
-              const keys = Object.keys(parseVal);
-              return <>{keys.join(', ')}</>;
-            },
-          },
-        },
-
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    Secret: {
-      name: 'Secret',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'type',
-          label: 'Type',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    ResourceQuota: {
-      name: 'ResourceQuota',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    LimitRange: {
-      name: 'LimitRange',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    HorizontalPodAutoscaler: {
-      name: 'HorizontalPodAutoscaler',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['spec.attribute', 'm'],
-        ['status.attribute', 'm'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Min Replicas',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let minReplicas = attribute?.minReplicas;
-              return <>{minReplicas}</>;
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Max Replicas',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let maxReplicas = attribute?.maxReplicas;
-              return <>{maxReplicas}</>;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Current Replicas',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let currentReplicas = attribute?.currentReplicas;
-              return <>{currentReplicas}</>;
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            sortThirdClickReset: true,
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    VerticalPodAutoscaler: {
-      name: 'VerticalPodAutoscaler',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    PodDisruptionBudget: {
-      name: 'PodDisruptionBudget',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['spec.attribute', 'm'],
-        ['status.attribute', 'm'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Min Available',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let minAvailable = attribute?.minAvailable;
-              return <>{minAvailable}</>;
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Max Available',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let maxAvailable = attribute?.maxAvailable;
-              return <>{maxAvailable}</>;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Current Healthy',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let currentHealthy = attribute?.currentHealthy;
-              return <>{currentHealthy}</>;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Desired Healthy',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let desiredtHealthy = attribute?.desiredtHealthy;
-              return <>{desiredtHealthy}</>;
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    PriorityClass: {
-      name: 'PriorityClass',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    RuntimeClass: {
-      name: 'RuntimeClass',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    Leases: {
-      name: 'Leases',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['spec.attribute', 'm'],
-        ['status.attribute', 'm'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Holder Identity',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let holderIdentity = attribute?.holderIdentity;
-              return <>{holderIdentity}</>;
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    MutatingWebhookConfiguration: {
-      name: 'MutatingWebhookConfiguration',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
+    ConfigMap: buildConfigMapColumns(args),
+    Secret: buildSecretColumns(args),
+    ResourceQuota: buildResourceQuotaColumns(args),
+    LimitRange: buildLimitRangeColumns(args),
+    HorizontalPodAutoscaler: buildHorizontalPodAutoscalerColumns(args),
+    VerticalPodAutoscaler: buildVerticalPodAutoscalerColumns(args),
+    PodDisruptionBudget: buildPodDisruptionBudgetColumns(args),
+    PriorityClass: buildPriorityClassColumns(args),
+    RuntimeClass: buildRuntimeClassColumns(args),
+    Leases: buildLeasesColumns(args),
+    MutatingWebhookConfiguration: buildMutatingWebhookConfigurationColumns(args),
   };
 };
