+++ ui/components/dashboard/resources/network/endpoint-slice-columns.tsx
@@ -1,0 +1,111 @@
+import React from 'react';
+import { timeAgo } from '../../../../utils/k8s-utils';
+import { getK8sContextFromClusterId } from '../../../../utils/multi-ctx';
+import { SINGLE_VIEW } from '../config';
+import { Title } from '../../view';
+import { TooltipWrappedConnectionChip } from '../../../connections/ConnectionChip';
+import { SortableTableCell } from '../sortable-table-cell';
+import { CONNECTION_KINDS } from '../../../../utils/Enum';
+import { FormatId } from '@/components/data-formatter';
+
+export const buildEndpointSliceColumns = ({
+  switchView,
+  meshSyncResources,
+  k8sConfig,
+  connectionMetadataState,
+  workloadType,
+  ping,
+}) => ({
+  name: 'EndpointSlice',
+  colViews: [
+    ['id', 'na'],
+    ['metadata.name', 'xs'],
+    ['apiVersion', 'na'],
+    ['metadata.namespace', 'm'],
+    ['cluster_id', 'xs'],
+    ['metadata.creationTimestamp', 'l'],
+  ],
+  columns: [
+    {
+      name: 'id',
+      label: 'ID',
+      options: {
+        display: false,
+        customBodyRender: (value) => <FormatId id={value} />,
+      },
+    },
+    {
+      name: 'metadata.name',
+      label: 'Name',
+      options: {
+        sort: false,
+        customBodyRender: function CustomBody(value, tableMeta) {
+          return (
+            <Title
+              onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
+              value={value}
+              kind={workloadType}
+            />
+          );
+        },
+      },
+    },
+    {
+      name: 'apiVersion',
+      label: 'API version',
+      options: {
+        sort: true,
+        sortThirdClickReset: true,
+        customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
+          return (
+            <SortableTableCell
+              index={index}
+              columnData={column}
+              columnMeta={columnMeta}
+              onSort={() => sortColumn(index)}
+            />
+          );
+        },
+      },
+    },
+    {
+      name: 'metadata.namespace',
+      label: 'Namespace',
+      options: {
+        sort: false,
+      },
+    },
+    {
+      name: 'cluster_id',
+      label: 'Cluster',
+      options: {
+        sort: true,
+        customBodyRender: function CustomBody(val) {
+          let context = getK8sContextFromClusterId(val, k8sConfig);
+          return (
+            <TooltipWrappedConnectionChip
+              title={context.name}
+              iconSrc={
+                connectionMetadataState
+                  ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
+                  : ''
+              }
+              handlePing={() => ping(context.name, context.server, context.connectionId)}
+            />
+          );
+        },
+      },
+    },
+    {
+      name: 'metadata.creationTimestamp',
+      label: 'Age',
+      options: {
+        sort: false,
+        customBodyRender: function CustomBody(value) {
+          let time = timeAgo(value);
+          return <>{time}</>;
+        },
+      },
+    },
+  ],
+});
