+++ ui/components/dashboard/resources/workloads/config.tsx
@@ -1,14 +1,12 @@
-import React from 'react';
-import { timeAgo } from '../../../../utils/k8s-utils';
-import { getK8sContextFromClusterId } from '../../../../utils/multi-ctx';
-import { SINGLE_VIEW } from '../config';
-import { Title } from '../../view';
-import { TooltipWrappedConnectionChip } from '../../../connections/ConnectionChip';
-import { ResizableCell } from '../../../../utils/utils';
 import useKubernetesHook from '@/utils/hooks/useKubernetesHook';
-import { DefaultTableCell, SortableTableCell } from '../sortable-table-cell';
-import { CONNECTION_KINDS } from '../../../../utils/Enum';
-import { FormatId } from '@/components/data-formatter';
+import { buildPodColumns } from './pod-columns';
+import { buildDeploymentColumns } from './deployment-columns';
+import { buildDaemonSetColumns } from './daemonset-columns';
+import { buildStatefulSetColumns } from './statefulset-columns';
+import { buildReplicaSetColumns } from './replicaset-columns';
+import { buildReplicationControllerColumns } from './replication-controller-columns';
+import { buildJobColumns } from './job-columns';
+import { buildCronJobColumns } from './cronjob-columns';
 
 export const WorkloadTableConfig = (
   switchView,
@@ -18,1255 +16,22 @@ export const WorkloadTableConfig = (
   workloadType,
 ) => {
   const ping = useKubernetesHook();
+  const args = {
+    switchView,
+    meshSyncResources,
+    k8sConfig,
+    connectionMetadataState,
+    workloadType,
+    ping,
+  };
   return {
-    Pod: {
-      name: 'Pod',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['status.attribute', 's'],
-        ['status.attribute', 's'],
-        ['status.attribute', 'm'],
-        ['metadata.namespace', 'm'],
-        ['spec.attribute', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-        ['status.attribute', 'm'],
-        ['spec.attribute', 'm'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            display: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            setCellProps: () => ({ style: { paddingRight: '0px', width: '0%' } }),
-            setCellHeaderProps: () => ({ style: { paddingRight: '0px' } }),
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Host IP',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let hostIP = attribute?.hostIP;
-              return <>{hostIP}</>;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Pod IP',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let podIP = attribute?.podIP;
-              return <>{podIP}</>;
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Restarts',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              const parsedStatus = JSON.parse(value);
-              const totalRestarts = parsedStatus?.containerStatuses?.reduce(
-                (sum, container) => sum + (container.restartCount || 0),
-                0,
-              );
-              return <>{totalRestarts || 0}</>;
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Containers',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              const parsedSpec = JSON.parse(value);
-              return <>{parsedSpec.containers.length}</>;
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Node',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let nodeName = attribute?.nodeName;
-              return (
-                <>
-                  <ResizableCell value={nodeName} />
-                </>
-              );
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <>
-                  <TooltipWrappedConnectionChip
-                    title={context.name}
-                    iconSrc={
-                      connectionMetadataState
-                        ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                        : ''
-                    }
-                    width="10.5rem"
-                    handlePing={() => ping(context.name, context.server, context.connectionId)}
-                  />
-                </>
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    Deployment: {
-      name: 'Deployment',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['status.attribute', 's'],
-        ['status.attribute', 's'],
-        ['spec.attribute', 's'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            display: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Pods',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              const parsedStatus = JSON.parse(val);
-              const pods =
-                parsedStatus?.replicas === undefined
-                  ? parsedStatus?.availableReplicas?.toString()
-                  : `${
-                      parsedStatus?.availableReplicas?.toString() ?? '0'
-                    } / ${parsedStatus?.replicas?.toString()}`;
-              return <>{pods}</>;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Replicas',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let replicas = attribute?.replicas;
-              return <>{replicas}</>;
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Restart Policy',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let template = attribute?.template;
-              let spec = template.spec;
-              let restartPolicy = spec.restartPolicy;
-              return <>{restartPolicy}</>;
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <>
-                  <TooltipWrappedConnectionChip
-                    title={context.name}
-                    iconSrc={
-                      connectionMetadataState
-                        ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                        : ''
-                    }
-                    handlePing={(event) => {
-                      event.preventDefault();
-                      ping(context.name, context.server, context.connectionId);
-                    }}
-                  />
-                </>
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    DaemonSet: {
-      name: 'DaemonSet',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['spec.attribute', 's'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            display: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Node Selector',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let template = attribute?.template;
-              let spec = template.spec;
-              let nodeSelector = spec.nodeSelector;
-              return <>{JSON.stringify(nodeSelector)}</>;
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    StatefulSet: {
-      name: 'StatefulSet',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['status.attribute', 's'],
-        ['status.attribute', 's'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            display: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Pods',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              const parsedStatus = JSON.parse(val);
-              const pods =
-                parsedStatus?.replicas === undefined
-                  ? parsedStatus?.availableReplicas?.toString()
-                  : `${
-                      parsedStatus?.availableReplicas?.toString() ?? '0'
-                    } / ${parsedStatus?.replicas?.toString()}`;
-              return <>{pods}</>;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Replicas',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let replicas = attribute?.replicas;
-              return <>{replicas}</>;
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    ReplicaSet: {
-      name: 'ReplicaSet',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['spec.attribute', 's'],
-        ['status.attribute', 's'],
-        ['status.attribute', 'm'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            display: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Desired Replicas',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let replicas = attribute?.replicas;
-              return <>{replicas}</>;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Current Replicas',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let replicas = attribute?.replicas;
-              return <>{replicas}</>;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Ready Replicas',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let readyReplicas = attribute?.readyReplicas;
-              return <>{readyReplicas}</>;
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            sortThirdClickReset: true,
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    ReplicationController: {
-      name: 'ReplicationController',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['spec.attribute', 's'],
-        ['status.attribute', 's'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            display: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Desired Replicas',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let replicas = attribute?.replicas;
-              return <>{replicas}</>;
-            },
-          },
-        },
-        {
-          name: 'status.attribute',
-          label: 'Current Replicas',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let replicas = attribute?.replicas;
-              return <>{replicas}</>;
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    Job: {
-      name: 'Job',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            display: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
-    CronJob: {
-      name: 'CronJob',
-      colViews: [
-        ['id', 'na'],
-        ['metadata.name', 'xs'],
-        ['apiVersion', 'na'],
-        ['spec.attribute', 's'],
-        ['spec.attribute', 's'],
-        ['metadata.namespace', 'm'],
-        ['cluster_id', 'xs'],
-        ['metadata.creationTimestamp', 'l'],
-      ],
-      columns: [
-        {
-          name: 'id',
-          label: 'ID',
-          options: {
-            display: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: (value) => <FormatId id={value} />,
-          },
-        },
-        {
-          name: 'metadata.name',
-          label: 'Name',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value, tableMeta) {
-              return (
-                <Title
-                  onClick={() => switchView(SINGLE_VIEW, meshSyncResources[tableMeta.rowIndex])}
-                  value={value}
-                  kind={workloadType}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'apiVersion',
-          label: 'API version',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Schedule',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let schedule = attribute?.schedule;
-              return <>{schedule}</>;
-            },
-          },
-        },
-        {
-          name: 'spec.attribute',
-          label: 'Suspend',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(val) {
-              let attribute = JSON.parse(val);
-              let suspend = attribute?.suspend;
-              return <>{suspend}</>;
-            },
-          },
-        },
-        {
-          name: 'metadata.namespace',
-          label: 'Namespace',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-          },
-        },
-        {
-          name: 'cluster_id',
-          label: 'Cluster',
-          options: {
-            sort: true,
-            sortThirdClickReset: true,
-            customHeadRender: function CustomHead({ index, ...column }, sortColumn, columnMeta) {
-              return (
-                <SortableTableCell
-                  index={index}
-                  columnData={column}
-                  columnMeta={columnMeta}
-                  onSort={() => sortColumn(index)}
-                />
-              );
-            },
-            customBodyRender: function CustomBody(val) {
-              let context = getK8sContextFromClusterId(val, k8sConfig);
-              return (
-                <TooltipWrappedConnectionChip
-                  title={context.name}
-                  iconSrc={
-                    connectionMetadataState
-                      ? connectionMetadataState[CONNECTION_KINDS.KUBERNETES]?.icon
-                      : ''
-                  }
-                  handlePing={() => ping(context.name, context.server, context.connectionId)}
-                />
-              );
-            },
-          },
-        },
-        {
-          name: 'metadata.creationTimestamp',
-          label: 'Age',
-          options: {
-            sort: false,
-            customHeadRender: function CustomHead({ ...column }) {
-              return <DefaultTableCell columnData={column} />;
-            },
-            customBodyRender: function CustomBody(value) {
-              let time = timeAgo(value);
-              return <>{time}</>;
-            },
-          },
-        },
-      ],
-    },
+    Pod: buildPodColumns(args),
+    Deployment: buildDeploymentColumns(args),
+    DaemonSet: buildDaemonSetColumns(args),
+    StatefulSet: buildStatefulSetColumns(args),
+    ReplicaSet: buildReplicaSetColumns(args),
+    ReplicationController: buildReplicationControllerColumns(args),
+    Job: buildJobColumns(args),
+    CronJob: buildCronJobColumns(args),
   };
 };
